Keep track of memory region types
Fixes #1. Data is now stored in a Memory object with tags describing the
type of data each region contains. Tags are assigned when appending new
data with the add_data, add_lit, or add_bss methods. Memory can be
indexed as if it were a bytearray but attempts to write to BSS regions
will fail.